# swagger-docs.json
{"openapi":"3.0.0","info":{"title":"Nest API Generate","description":"API built with Nest","version":"0.0.1","contact":{}},"tags":[],"servers":[{"url":"http://localhost:3333"}],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"StoreUserDTO":{"type":"object","properties":{"name":{"type":"string"},"lastname":{"type":"string"},"email":{"type":"string"},"photoURL":{"type":"string"},"password":{"type":"string"},"emailIsConfirmed":{"type":"boolean","default":false},"whatsapp":{"type":"string"}},"required":["name","lastname","email","photoURL","password"]},"User":{"type":"object","properties":{"firebaseId":{"type":"string"},"name":{"type":"string"},"photoURL":{"type":"string"},"lastname":{"type":"string"},"roles":{"type":"array","items":{"type":"string"}},"email":{"type":"string"},"isDeleted":{"type":"boolean","default":false},"createdAt":{"format":"date-time","type":"string"}},"required":["firebaseId","name","photoURL","lastname","roles","email","createdAt"]},"UpdateUserDTO":{"type":"object","properties":{"name":{"type":"string"},"lastname":{"type":"string"},"email":{"type":"string"},"photoURL":{"type":"string"},"password":{"type":"string"},"emailIsConfirmed":{"type":"boolean","default":false},"whatsapp":{"type":"string"}},"required":["name","lastname","email","photoURL","password"]},"PatchUserDTO":{"type":"object","properties":{"name":{"type":"string"},"lastname":{"type":"string"},"email":{"type":"string"},"photoURL":{"type":"string"},"password":{"type":"string"},"emailIsConfirmed":{"type":"boolean","default":false},"whatsapp":{"type":"string"}}}}},"paths":{"/authentications/sign-up-user":{"post":{"operationId":"AuthenticationsController_signUpUser","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/StoreUserDTO"}}}},"responses":{"201":{"description":"Produtor cadastrado com sucesso."}},"tags":["authentications"]}},"/authentications/sign-up-admin":{"post":{"operationId":"AuthenticationsController_signUpAdmin","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/StoreUserDTO"}}}},"responses":{"201":{"description":"Administrador cadastrado com sucesso."}},"tags":["authentications"]}},"/authentications/sign-google-user":{"post":{"operationId":"AuthenticationsController_SignClientWithGoogle","parameters":[],"responses":{"201":{"description":""}},"tags":["authentications"]}},"/users":{"get":{"operationId":"UsersController_getAll","parameters":[{"name":"skip","required":false,"in":"query","schema":{"type":"number"}},{"name":"limit","required":false,"in":"query","schema":{"type":"number"}},{"name":"roles","required":false,"in":"query","schema":{"type":"string"}},{"name":"email","required":false,"in":"query","schema":{"type":"string"}},{"name":"zoopId","required":false,"in":"query","schema":{"type":"string"}},{"name":"fromDate","required":false,"in":"query","schema":{"type":"date"}},{"name":"toDate","required":false,"in":"query","schema":{"type":"date"}}],"responses":{"200":{"description":"Usuários obtidas com sucesso.","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/PaginatedResponse"},{"properties":{"pagination":{"$ref":"#/components/schemas/PaginationInfos"},"results":{"type":"array","items":{"$ref":"#/components/schemas/User"}}}}]}}}}},"tags":["users"]}},"/users/{id}":{"get":{"operationId":"UsersController_getOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Usuário obtida com sucesso.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"tags":["users"]},"put":{"operationId":"UsersController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDTO"}}}},"responses":{"200":{"description":"Usuário editada com sucesso.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"tags":["users"]},"patch":{"operationId":"UsersController_patch","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PatchUserDTO"}}}},"responses":{"200":{"description":"Usuário editada com sucesso.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"tags":["users"]},"delete":{"operationId":"UsersController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Usuário removida com sucesso."}},"tags":["users"]}},"/users/{id}/thumbnail/upload":{"post":{"operationId":"UsersController_updateThumbnail","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"tags":["users"]}},"/users/{id}/inactivate":{"post":{"operationId":"UsersController_inactivateUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"tags":["users"]}},"/users/{id}/activate":{"post":{"operationId":"UsersController_activateUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"tags":["users"]}}}}